Fixing various migration issues, migrating from bigint to big serial

+++ api/src/main/resources/db/migration/V1_12__CREATE_FacilityTimetable_TABLE.sql
@@ -1,8 +1,8 @@
 create table sports_centre_management.facility_timetable
 (
-    id         int8 not null,
+    id         bigserial not null,
     name       varchar(255),
-    centre_id  int8 not null,
+    centre_id  int8                  not null,
     created_at timestamp,
     updated_at timestamp,
     primary key (id)
